\name{getDBVersion}
\alias{getDBVersion}

\title{HFA-DB version}

\description{
Print current version of the HFA-DB.
}

\usage{
getDBVersion()
}

\author{\email{brechtdv@gmail.com}}

\seealso{
\code{\link{getHFADB}} to get the current HFA database URL\cr
\code{\link{setHFADB}} to set the HFA database URL
}