\name{getHFADB}
\alias{getHFADB}

\title{Get current HFA database URL}

\description{
Print the URL of the current HFA database.
}

\usage{
getHFADB()
}

\author{\email{brechtdv@gmail.com}}

\seealso{
\code{\link{setHFADB}} to set HFA database URL\cr
\code{\link{getDBVersion}} to get the current version of the HFA database
}