\name{getYears}
\alias{getYears}

\title{Get Years}

\description{
Extract the years for which data are available for a given country.
}

\usage{
getYears(country)
}

\arguments{
  \item{country}{ID (character) of the country; only one ID is allowed.}
}

\author{\email{brechtdv@gmail.com}}

\value{
Numeric vector.
}

\seealso{
\code{\link{getCountries}} to extract the set of available countries
}